requirements.txt
torch==1.13.1
transformers[torch]==4.26.0
sentencepiece
flask
boto3
selenium
beautifulsoup4
pydantic
langchain==0.0.171
requests
protobuf
numpy
pandas
scipy
openai
python-dotenv
langchain-community
